remade

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -5,6 +5,7 @@ import jm.task.core.jdbc.util.Util;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class UserDaoHibernateImpl implements UserDao {
@@ -43,11 +44,15 @@ public class UserDaoHibernateImpl implements UserDao {
     @Override
     public void saveUser(String name, String lastName, byte age) {
         try (Session session = Util.getSessionFactory().openSession()) {
-            session.beginTransaction();
-            User user = new User(name, lastName, age);
-            session.save(user);
-            System.out.println(user);
-            session.getTransaction().commit();
+           try {
+               session.beginTransaction();
+               User user = new User(name, lastName, age);
+               session.save(user);
+               System.out.println(user);
+               session.getTransaction().commit();
+           } catch (HibernateException e) {
+               session.getTransaction().rollback();
+           }
         } catch (HibernateException e) {
             e.printStackTrace();
         }
@@ -56,9 +61,15 @@ public class UserDaoHibernateImpl implements UserDao {
     @Override
     public void removeUserById(long id) {
         try (Session session = Util.getSessionFactory().openSession()) {
-            session.beginTransaction();
-            session.delete(session.get(User.class, id));
-            session.getTransaction().commit();
+            try {
+                session.beginTransaction();
+                if (session.get(User.class, id) != null) {
+                    session.delete(session.get(User.class, id));
+                    session.getTransaction().commit();
+                }
+            } catch (HibernateException e) {
+                session.getTransaction().rollback();
+            }
         } catch (HibernateException e) {
             e.printStackTrace();
         }
@@ -66,14 +77,14 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public List<User> getAllUsers() {
+        List<User> users = new ArrayList<User>();
         try (Session session = Util.getSessionFactory().openSession()) {
-            List<User> users = session.createQuery("from User").list();
+            users = session.createQuery("from User").list();
             users.forEach(System.out::println);
-            return users;
         } catch (HibernateException e) {
             e.printStackTrace();
         }
-        return null;
+        return users;
     }
 
     @Override
@@ -85,10 +96,5 @@ public class UserDaoHibernateImpl implements UserDao {
         } catch (HibernateException e) {
             e.printStackTrace();
         }
-    }
-
-    @Override
-    public void closeConnection() {
-        Util.getSessionFactory().close();
     }
 }
